src/maintenance_server/files/home/tasks/vaultbackup/requirements.txt
pycryptodome
pyyaml
hvac
boto3
botocore
